api/app/requirements.txt
fastapi
uvicorn
paho-mqtt
python-multipart        # si piensas recibir ficheros
feedparser
pyyaml
openai
